requirements.txt
streamlit
gtts
cohere
Pillow
tensorflow
numpy